+++ crates/reth/evm/Cargo.toml
@@ -12,6 +12,8 @@ strata-crypto.workspace = true
 strata-primitives.workspace = true
 
 alloy-sol-types.workspace = true
+reth.workspace = true
+reth-evm.workspace = true
 reth-primitives.workspace = true
 revm.workspace = true
 revm-primitives.workspace = true
