+++ crates/reth/exex/Cargo.toml
@@ -28,9 +28,9 @@ reth-trie.workspace = true
 reth-trie-common = { workspace = true, features = ["eip1186"] }
 revm-primitives.workspace = true
 rockbound.workspace = true
-rsp-client-executor = { git = "https://github.com/succinctlabs/rsp", rev = "3703a82027157a653327c95cf2e507d05bde174f" }
-rsp-mpt = { git = "https://github.com/succinctlabs/rsp", rev = "3703a82027157a653327c95cf2e507d05bde174f" }
-rsp-primitives = { git = "https://github.com/succinctlabs/rsp", rev = "3703a82027157a653327c95cf2e507d05bde174f" }
+rsp-client-executor = { git = "https://github.com/alpenlabs/rsp", rev = "a386ada51586761058eee16acdab1561f06947c5" }
+rsp-mpt = { git = "https://github.com/alpenlabs/rsp", rev = "a386ada51586761058eee16acdab1561f06947c5" }
+rsp-primitives = { git = "https://github.com/alpenlabs/rsp", rev = "a386ada51586761058eee16acdab1561f06947c5" }
 serde_json.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
